Massive renderer / rendering context refactor

+++ Framework/RenderingDefines.hpp
@@ -11,7 +11,7 @@
 
 namespace dm {
 
-constexpr size_t MAX_FRAME_DRAWS = 2;
+constexpr int MAX_FRAME_DRAWS = 2;
 
 const std::vector<const char*> deviceExtensions = {
 	VK_KHR_SWAPCHAIN_EXTENSION_NAME
